requirements.txt
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.0
